minor fixes

+++ src/graphics_segment.cpp
@@ -4,14 +4,16 @@
 namespace runebound::graphics {
 void Segment::render(SDL_Renderer *renderer, SDL_Color col, int size) const {
     thickLineRGBA(
-        renderer, m_s.x(), m_s.y(), m_f.x(), m_f.y(), 2 * size + 1, col.r,
-        col.g, col.b, col.a
+        renderer, (short)m_s.x(), (short)m_s.y(), (short)m_f.x(),
+        (short)m_f.y(), 2 * size + 1, col.r, col.g, col.b, col.a
     );
     filledCircleRGBA(
-        renderer, m_s.x(), m_s.y(), size, col.r, col.g, col.b, col.a
+        renderer, (short)m_s.x(), (short)m_s.y(), (short)size, col.r, col.g,
+        col.b, col.a
     );
     filledCircleRGBA(
-        renderer, m_f.x(), m_f.y(), size, col.r, col.g, col.b, col.a
+        renderer, (short)m_f.x(), (short)m_f.y(), (short)size, col.r, col.g,
+        col.b, col.a
     );
 }
 
@@ -19,11 +21,13 @@ void Segment::render(SDL_Renderer *renderer, SDL_Color col, int size) const {
 void Segment::half_render(SDL_Renderer *renderer, SDL_Color col, int size)
     const {
     thickLineRGBA(
-        renderer, m_s.x(), m_s.y(), (m_s.x() + m_f.x()) / 2,
-        (m_s.y() + m_f.y()) / 2, 2 * size + 1, col.r, col.g, col.b, col.a
+        renderer, (short)m_s.x(), (short)m_s.y(),
+        (short)((m_s.x() + m_f.x()) / 2), (short)((m_s.y() + m_f.y()) / 2),
+        2 * size + 1, col.r, col.g, col.b, col.a
     );
     filledCircleRGBA(
-        renderer, m_s.x(), m_s.y(), size, col.r, col.g, col.b, col.a
+        renderer, (short)m_s.x(), (short)m_s.y(), (short)size, col.r, col.g,
+        col.b, col.a
     );
 }
 }  // namespace runebound::graphics
